+++ cliquet/resource/__init__.py
@@ -977,6 +977,15 @@ class ProtectedResource(BaseResource):
                 self.context.get_permission_object_id,
                 self.request)
 
+    def get_parent_id(self, request):
+        """Unlike :class:`BaseResource`, records are not isolated by user.
+
+        See https://github.com/mozilla-services/cliquet/issues/549
+
+        :returns: A constant empty value.
+        """
+        return ''
+
     def _extract_filters(self, queryparams=None):
         """Override default filters extraction from QueryString to allow
         partial collection of records.
@@ -1019,13 +1028,13 @@ class ProtectedResource(BaseResource):
 
         return annotated
 
-    def postprocess(self, result, action=ACTIONS.READ):
+    def postprocess(self, result, **kwargs):
         """Add ``permissions`` attribute in response body.
 
         In the protocol, it was decided that ``permissions`` would reside
         outside the ``data`` attribute.
         """
-        result = super(ProtectedResource, self).postprocess(result)
+        result = super(ProtectedResource, self).postprocess(result, **kwargs)
         if isinstance(result['data'], list):
             # collection endpoint.
             return result