+++ cliquet/tests/test_cache.py
@@ -2,6 +2,7 @@ import mock
 import time
 
 import redis
+from pyramid import testing
 
 from cliquet.utils import psycopg2
 from cliquet.storage import exceptions
@@ -82,6 +83,15 @@ class BaseTestCache(object):
         with mock.patch('cliquet.cache.random.random', return_value=0.4):
             self.assertTrue(ping(self.request))
 
+    def test_ping_logs_error_if_unavailable(self):
+        self.client_error_patcher.start()
+        ping = heartbeat(self.cache)
+
+        with mock.patch('cliquet.cache.logger.exception') as exc_handler:
+            self.assertFalse(ping(self.request))
+
+        self.assertTrue(exc_handler.called)
+
     def test_set_adds_the_record(self):
         stored = 'toto'
         self.cache.set('foobar', stored)
@@ -141,6 +151,9 @@ class MemoryCacheTest(BaseTestCache, unittest.TestCase):
     def test_ping_returns_false_if_unavailable(self):
         pass
 
+    def test_ping_logs_error_if_unavailable(self):
+        pass
+
 
 class RedisCacheTest(BaseTestCache, unittest.TestCase):
     backend = redis_backend
@@ -155,6 +168,14 @@ class RedisCacheTest(BaseTestCache, unittest.TestCase):
             self.cache._client,
             'execute_command',
             side_effect=redis.RedisError)
+
+    def test_config_is_taken_in_account(self):
+        config = testing.setUp(settings=self.settings)
+        config.add_settings({'cache_url': 'redis://:secret@peer.loc:4444/7'})
+        backend = self.backend.load_from_config(config)
+        self.assertDictEqual(
+            backend._client.connection_pool.connection_kwargs,
+            {'host': 'peer.loc', 'password': 'secret', 'db': 7, 'port': 4444})
 
 
 @skip_if_no_postgresql